requirements.txt
numpy
pandas
python-dotenv
snowflake-connector-python
boto3
s3fs